Improved Performance We made the creation of channels faster since now we don't need to search for the cateogry each time We don't delete channels if there are no channels in system

+++ Siggraph Discord.py
@@ -36,9 +36,10 @@ async def ping(ctx):
 
 @bot.command(name='purge', description='delete every channel here in this system', brief='DELETE EVERYTHING')
 async def purge(ctx):
-    for channel in bot.get_all_channels():
-        # print(channel.name)
-        await channel.delete()
+    if len(list(bot.get_all_channels())) > 0:
+        for channel in bot.get_all_channels():
+            print(channel.name)
+            await channel.delete()
     await ctx.send('All channels and categories are gone!!!')
 
 
@@ -46,22 +47,29 @@ async def purge(ctx):
 @bot.command(name='create_from_CSV', description='create channels and categories from CSV', brief='starts the new world ')
 async def createFromCSV(ctx):
     df = pd.read_csv("..\s2020_sessions_Test.csv")
-
+    categories = {}
     for event_type in df["Event Types"].unique():
-        await bot.guilds[0].create_category(event_type)
+        category = await bot.guilds[0].create_category(event_type)
+        categories[event_type] = category
     await ctx.send('created all the categories!')
 
     df["Reduced_sessionTitle"] = df['Session Title'].str.strip().str[:20]
 
     for index, row in df.iterrows():
         # print(row['Reduced_sessionTitle'], row['Event Types'])
-        category = discord.utils.get(
-            bot.guilds[0].categories, name=row['Event Types'])
+        # category = discord.utils.get(
+        #     bot.guilds[0].categories, name=row['Event Types'])
         # We can't have more than 50 channels in category
-        if len(category.channels)<50:
-            await bot.guilds[0].create_text_channel(row['Reduced_sessionTitle'], category=category)
+        if len(categories[row['Event Types']].channels) < 50:
+            await bot.guilds[0].create_text_channel(row['Reduced_sessionTitle'], category=categories[row['Event Types']])
     await ctx.send('All channels and categories are created from CSV!!!')
 
+
+# TODO: The Command below is not working. Sending multiple delete messages
+@bot.command(name='reset', description='delete everything and create again from a csv', brief='restart the world', hidden=True)
+async def createFromCSV(ctx):
+    await purge(ctx)
+    await createFromCSV(ctx)
 
 # Commands don't work when this is set
 # @bot.event
